I have added the createCompra use case.

+++ src/domain/puntos-leal/puntos-leal.entity.go
@@ -4,5 +4,5 @@ type PuntosLeal struct {
 	ID             int16
 	Comercio       int16
 	Usuario        int16
-	CantidadPuntos int32
+	CantidadPuntos float64
 }
