+++ src/domain/usuarios/usuario.entity.go
@@ -3,6 +3,6 @@ package domain
 type Usuario struct {
 	ID             int16
 	PuntosLeal     int16
-	PuntosCashback int32
+	PuntosCashback float64
 	Nombre         string
 }
